Reorganized file layout and updated imports
