+++ frontend/src/editor/TranscriptionTextarea.jsx
