+++ frontend/src/viewer/GammaFilterForm.jsx
