+++ frontend/src/viewer/KeyboardShortcutRow.jsx
